+++ firefoxThemes/modules/utils/states.js
@@ -1,0 +1,12 @@
+//Enums for the different Mixed option states
+const mixedStates = {
+  NONE: 0,
+  INITIAL: 1,
+  RESET: 2
+};
+/*Enums for background types*/
+const backgroundTypes = {
+  PRIMARY: 0,
+  SECONDARY: 1
+};
+export {mixedStates,backgroundTypes};