+++ firefoxThemes/popup/popup.js
@@ -1,3 +1,8 @@
+import {svgToPng, buildSVG} from "../modules/utils/themes/logo.js";
+import {getRandomThemeId} from "../modules/utils/random.js";
+import {mixedStates, backgroundTypes} from "../modules/utils/states.js";
+import {optionsPage} from "../modules/utils/browser.js";
+
 /*Global Variables*/
 const selectTag = document.querySelector("select");
 const backgroundSwitch = document.querySelector("#backgroundType");
@@ -6,33 +11,23 @@ const darkModeSwitch = document.querySelector("#darkMode");
 const dokiHeart = document.querySelector("#doki_heart");
 const root = document.querySelector(":root");
 
-//Enum for the different Mixed option states
-const mixedStates = {
-  NONE: 0,
-  INITIAL: 1
-};
-
-const backgroundTypes = {
-  PRIMARY: 0,
-  SECONDARY: 1
-};
 
 /*Set color of popup menu based on theme*/
 function setCSS(chosenTheme) {
   if (!chosenTheme) return;
-  const {colors, information} = chosenTheme.definition;
-  root.style.setProperty('--switch-shadow-color',information.dark ? 'white' : 'black');
-  root.style.setProperty('--doki-shadow',information.dark ? '11px #fff' : '11px #000');
-  root.style.setProperty('--info-foreground-color',colors.infoForeground);
-  root.style.setProperty('--header-color',colors.headerColor);
-  root.style.setProperty('--line-number-color',colors.lineNumberColor);
-  root.style.setProperty('--selection-inactive-color',colors.selectionInactive);
-  root.style.setProperty('--primary-accent-color',colors.accentColor);
-  root.style.setProperty('--secondary-accent-color',colors.accentColor+'44');
-  root.style.setProperty('--button-color',colors.buttonColor);
-  root.style.setProperty('--button-font-color',colors.buttonFont);
-  root.style.setProperty('--base-background-color',colors.baseBackground);
-
+  themeDokiLogo(chosenTheme);
+  const {colors} = chosenTheme.definition;
+  root.style.setProperty('--switch-shadow-color', colors.accentColor);
+  root.style.setProperty('--doki-shadow', colors.accentColor);
+  root.style.setProperty('--info-foreground-color', colors.infoForeground);
+  root.style.setProperty('--header-color', colors.headerColor);
+  root.style.setProperty('--line-number-color', colors.lineNumberColor);
+  root.style.setProperty('--selection-inactive-color', colors.selectionInactive);
+  root.style.setProperty('--primary-accent-color', colors.accentColor);
+  root.style.setProperty('--secondary-accent-color', colors.accentColorTransparent);
+  root.style.setProperty('--button-color', colors.buttonColor);
+  root.style.setProperty('--button-font-color', colors.buttonFont);
+  root.style.setProperty('--base-background-color', colors.baseBackground);
 }
 
 const setBackground = async () => {
@@ -90,18 +85,26 @@ const setOpposingTheme = async () => {
   }
 }
 
+function themeDokiLogo(currentTheme) {
+  const searchOptions = {width: 75, height: 75};
+  const searchSVG = buildSVG(currentTheme, searchOptions);
+  svgToPng(searchSVG, searchOptions, (imgData) => {
+    dokiHeart.src = imgData;
+  });
+}
+
 /*EVENT: Retrieve the selected waifu.
 Afterwards, send the chosen waifu to the background script.*/
 function setTheme(e) {
-  browser.storage.local.get(["darkMode", "waifuThemes"])
+  browser.storage.local.get(["darkMode", "waifuThemes", "mixedTabs"])
     .then((storage) => {
       const chosenThemeName = e.target.value;
-      const currentMix = chosenThemeName === "mixed" ? mixedStates.INITIAL : mixedStates.NONE;
+      const currentMix = getMixState(chosenThemeName, storage.mixedTabs);
       let chosenThemeId;
       if (currentMix === mixedStates.NONE) {
         let themes;
         if (chosenThemeName === "random") {
-          chosenThemeId = getRandomTheme(storage.waifuThemes.themes);
+          chosenThemeId = getRandomThemeId(storage.waifuThemes.themes);
           const chosenRandom = storage.waifuThemes.themes[chosenThemeId];
           themes = Object.values(storage.waifuThemes.themes)
             .filter(dokiTheme => (
@@ -109,7 +112,7 @@ function setTheme(e) {
                 || dokiTheme.displayName === chosenRandom.displayName)
               && dokiTheme.group === chosenRandom.group
             ));
-        }else{
+        } else {
           themes = Object.values(storage.waifuThemes.themes)
             .filter(dokiTheme => (
               dokiTheme.displayName === chosenThemeName ||
@@ -133,31 +136,32 @@ function setTheme(e) {
             setCSS(usableTheme);
             chosenThemeId = usableTheme.id;
           }
-        }else{
-          setCSS(storage.waifuThemes.themes[chosenThemeId]);
-          selectTag.value = 'none';
+        } else {
+          if (chosenThemeId) {
+            setCSS(storage.waifuThemes.themes[chosenThemeId]);
+          }
+          selectTag.value = 'none';// Reset option back to 'choose a waifu'
         }
+        setHasSecondary(storage.waifuThemes.themes[chosenThemeId]);
+      } else {
+        let themes = storage.waifuThemes.themes;
+        chosenThemeId = getRandomThemeId(themes);
+        setDarkMode(true);
+        setHasSecondary(false);
+        setCSS(themes[chosenThemeId]);
+        selectTag.value = 'none';// Reset option back to 'choose a waifu'
       }
-      setHasSecondary(storage.waifuThemes.themes[chosenThemeId]);
-      browser.runtime.sendMessage({currentThemeId: chosenThemeId || 'mixed', mixState: currentMix});
+      browser.runtime.sendMessage({currentThemeId: chosenThemeId, mixState: currentMix});
     });
 }
 
-/*Selects a waifu at random*/
-function getRandomTheme(themes) {
-  themes = Object.keys(themes);
-  const randNum = getRandomNumber(0, themes.length);
-  return themes[randNum];
-}
-
-/*Retrieves a random number from min(inclusive) to max(exclusive)*/
-function getRandomNumber(min, max) {
-  return Math.floor(Math.random() * (max - min)) + min;
-}
-
-/*Navigate to options page*/
-function optionsPage() {
-  browser.runtime.openOptionsPage();
+function getMixState(name, mixTabs) {
+  if (name === "mixed" && mixTabs) {
+    return mixedStates.RESET;
+  } else if (name === "mixed") {
+    return mixedStates.INITIAL;
+  }
+  return mixedStates.NONE;
 }
 
 /*Initializes checkbox switches*/
@@ -171,7 +175,7 @@ function prepareSwitches(storage) {
     darkModeSwitch.checked = false;
     darkModeSwitch.disabled = true;
   }
-  showSearchSwitch.checked = !!storage.showWidget;
+  showSearchSwitch.checked = storage.showWidget === undefined || storage.showWidget;
 }
 
 /*Setup Waifu Choices for the popup menu
@@ -228,9 +232,12 @@ function initChoice() {
           const themes = storage.waifuThemes.themes;
           if (activeTab && storage.mixedTabs) {
             const tabThemeId = storage.mixedTabs.get(activeTab.id);
-            setCSS(themes[tabThemeId]);
-            selectTag.options.selectedIndex =
-              selectTag.options.namedItem(tabThemeId).index;
+            let currentTheme = themes[tabThemeId];
+            if (currentTheme) {
+              setCSS(currentTheme);
+              let option = selectTag.options.namedItem(currentTheme.displayName);
+              selectTag.options.selectedIndex = option ? option.index : 0;
+            }
           } else if (activeTab) {
             setCSS(themes[storage.currentThemeId]);
           }
@@ -245,3 +252,5 @@ backgroundSwitch.addEventListener("change", setBackground, true);
 showSearchSwitch.addEventListener("change", setHideWidget, true);
 darkModeSwitch.addEventListener("change", setOpposingTheme, true);
 dokiHeart.addEventListener('click', optionsPage, true);
+/*Exports*/
+export {getRandomThemeId};